Refactor directory hierarchy
