+++ README.md
@@ -21,4 +21,4 @@ cmake --build build
 cd build && ctest
 ```
 
-When compiling the main program, make sure to have the flags `-O3` and `-fopenmp`.
+When compiling the main program, make sure to have the flags `-O3` and `-fopenmp` (this is done automatically with the Makefile).
