+++ examples/README.md
@@ -1,0 +1,3 @@
+# Examples
+
+Example files of how to use PyGrad. To compile them, use the `make` utility. You'll need to download the [Iris Dataset](https://www.kaggle.com/uciml/iris/version/2)
